app/requirements.txt
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
Werkzeug==1.0.1
flask_restx==0.2.0
flask.json==0.3.4
flask_cors==3.0.10
sqlalchemy==1.3.23
flask_jwt_extended==4.0.2
boto3==1.17.17
Pillow==8.1.0
mysql-connector==2.2.9
